Lagt til presentasjon av resultat
