+++ web/src/test/resources/52AF.dat
@@ -1,0 +1,37 @@
+6676000170201001000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201002000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201003000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+667600 160201004000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000730201005000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201006000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201006000101202041119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201007000000000011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201008000101202001119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201009000101202016119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201010000101202011319761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+66760001702010110001012020111    1 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201012000101202011119767 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201013000101202011119764 1133  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201014000101202011119761  113  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201015000101202011119761 1 13  1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201016000101202011119761 111   1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201017000101202011119761 112  11       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201018000101202011119761 112 1 1       0302202021                 12122222220302              003000003005      00300521          10120052020
+6676000170201020100101202011119761 1113  1               21                 12122222220302              003000003005      00300521          10120052020
+6676000170201021000101202011119761 1113  1       0312201621                 12122222220302              003000003005      00300521          10120052020
+6676000170201022000101202011119761 1113  1       030220202                  12122222220302              003000003005      00300521          10120052020
+6676000170201023000101202011119761 1113  1       0302202021                  2122222220302              003000003005      00300521          10120052020
+6676000170201024000101202011119761 1113  1       0302202021                 12 22222220302              003000003005      00300521          10120052020
+6676000170201025000101202011119761 1113  1       0302202021                 1222222222                  003000003005      00300521          10120052020
+6676000170201026000101202011119761 1113  1       0302202021                 121222222203                003000003005      00300521          10120052020
+6676000170201027000101202011119761 1113  1       0302202021                 12122222220302              000000003005      00300521          10120052020
+6676000170201028000101202011119761 1113  1       0302202021                 12122222220302              003000000005      00300521          10120052020
+6676000170201029000101202011119761 1113  1       0302202021                 12122222220302              003000003000      00300521          10120052020
+6676000170201030000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00000521          10120052020
+6676000170201031000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300021          10120052020
+6676000170201032000101202011119761 1113  1       0302202021                 12122222220302              003000003005      0030052           10120052020
+6676000170201034000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521           0120052020
+6676000170201033000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          1  20052020
+6676000170201035000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          101       s
+6676000170201036000101202011119761 1113  1       0302202021                 12122222220302              003000003005      00300521          10120052018
+6676000170201037000101202011119761 1113  1       0302202021                 1 122222220302              003000003005      00300521          10120052020
